requirements.txt
requests
pandas
datetime
ccxt 
ta 
numpy 
fredapi 
pytz 
yfinance 
bs4